lib: update ce-mask to uint64_t
lib/route/link.c already defines 32 attributes which fills the current
uint32_t used for ce_mask. To accommodate more attributes the mask needs
to be expanded. This patch updates the definition to uint64_t.

The nl_object_diff API is maintained for ABI with existing users. A new
nl_object_diff64 API is added for the expanded attribute list. The MSB
of the 32-bit API is used to indicate if higher order attributes had a
mismatch. (Suggested by Thomas).

Note that LINK_ATTR_LINK_NETNSID changes. But since the attribute flags
are not public API it shouldn't be a problem.

http://lists.infradead.org/pipermail/libnl/2015-December/002078.html
http://lists.infradead.org/pipermail/libnl/2015-December/002083.html

+++ include/netlink-private/object-api.h
@@ -126,6 +126,8 @@ extern "C" {
  * #define MY_ATTR_FOO		(1<<0)
  * #define MY_ATTR_BAR		(1<<1)
  *
+ * // Bit 31 for attributes is reserved for 32-bit API.
+ *
  * // When assigning an optional attribute to the object, make sure
  * // to mark its availability.
  * my_obj->foo = 123123;
@@ -189,7 +191,7 @@ extern "C" {
 	struct nl_list_head	ce_list;	\
 	int			ce_msgtype;	\
 	int			ce_flags;	\
-	uint32_t		ce_mask;
+	uint64_t		ce_mask;
 
 struct nl_object
 {
@@ -258,7 +260,7 @@ struct nl_object
  * @endcode
  */
 #define ATTR_DIFF(LIST, ATTR, A, B, EXPR) \
-({	int diff = 0; \
+({	uint64_t diff = 0; \
 	if (((LIST) & (ATTR)) && ATTR_MISMATCH(A, B, ATTR, EXPR)) \
 		diff = ATTR; \
 	diff; })
@@ -333,8 +335,8 @@ struct nl_object_ops
 	 * The function must return a bitmask with the relevant bit
 	 * set for each attribute that mismatches.
 	 */
-	int   (*oo_compare)(struct nl_object *, struct nl_object *,
-			    uint32_t, int);
+	uint64_t (*oo_compare)(struct nl_object *, struct nl_object *,
+			       uint64_t, int);
 
 
 	/**
